Refactor ToolItem component structure and update imports
Reorganized the ToolItem component into its own dedicated folder (`ToolItem/ToolItem.tsx`) for better modularity and clarity. Updated all related imports and mocks to reflect the new directory structure. There are no functional changes in this update.
